test(ai): run live model evaluations through Vitest
Collect opt-in AI evaluations from a shared directory with retries and credentials loaded before client imports, while excluding them from deterministic unit runs.
